harden VM and extend the language
Bug fixes (memory safety / correctness):
- compiler: fix out-of-bounds write on >256 locals/registers. error() was
  non-fatal but callers still indexed past the fixed locals[256]/reg_class_hints
  arrays; add next_local() guard and clamp alloc_reg().
- arena: free large (>640B) allocations on destroy (they were leaked). Large
  blocks now carry a LargeHeader and are threaded on an intrusive list; O(1)
  register/unregister/realloc.
- vm: OP_MOD used a 32-bit divisor (wrong result for divisors > 2^31); use int64.
- parser: add recursion-depth guard (kMaxParseDepth) in parse_precedence and
  statement() to prevent stack-overflow segfaults on deeply nested source.
- compile: pause the GC for the whole compile so a collection triggered mid-emit
  cannot free the ObjFunc being built (double-free/UAF on large functions).

Language / semantics:
- add %= (modulo-assign).
- ** now binds tighter than a prefix unary: -2 ** 2 == -(2**2) == -4.
- integer literals parse via strtoull: INT64_MIN literal is writable,
  0xFFFFFFFFFFFFFFFF == -1, and out-of-range decimals error instead of clamping.
- detect undefined variables at compile time (read of a name never
  defined/assigned/imported/built-in), with zero runtime cost.
- add pcall(fn, args...) for protected calls (Lua-style error handling); no
  try/catch, no longjmp, propagates cooperatively through native module frames.

Tooling / docs:
- add ZEN_ARENA_BYPASS build flag to A/B the arena vs plain malloc.
- fix null -> nil in tutorial_01_variaveis.zen; document pcall, %=, precedence
  and the updated operator table in syntax.md.

+++ libzen/src/vm_dispatch.cpp
@@ -667,7 +667,7 @@ namespace zen
             {
                 if (vb.type == VAL_INT && vc.type == VAL_INT)
                 {
-                    int32_t divisor = vc.as.integer;
+                    int64_t divisor = vc.as.integer;
                     if (divisor == 0)
                         R[ZEN_A(i)] = val_int(0);
                     else
